fix query

+++ src/main/java/api/m2/movements/records/balance/BalanceByCategoryRecord.java
@@ -3,8 +3,8 @@ package api.m2.movements.records.balance;
 import java.math.BigDecimal;
 
 public record BalanceByCategoryRecord(String category,
-                                      Integer year,
-                                      Integer month,
+                                      Long year,
+                                      Long month,
                                       String currencySymbol,
                                       BigDecimal total) {
 }
